Add install_requires to setup.py for pip install
Currently, when you install the package via "pip install
ohmysportsfeedspy", it doesn't install any of the requirements.
By moving the requirements into the setup.py, this problem will be
fixed. These versions should be as loose as possible for dependency
resolution.

Using the "." in requirements.txt will enable the existing make commands
to work.